+++ packages/blinkid-wasm/src/wasm-module.ts
@@ -2,11 +2,7 @@
  * Copyright (c) 2026 Microblink Ltd. All rights reserved.
  */
 
-import type { EmscriptenModule } from "@microblink/wasm-common";
-import {
-  LicenseUnlockResult,
-  ServerPermissionSubmitError,
-} from "@microblink/wasm-common";
+import type { EmscriptenModule, WasmBindings } from "@microblink/wasm-common";
 import { BlinkIdScanningSession, BlinkIdSessionSettings } from "./session";
 
 /**
@@ -21,30 +17,7 @@ export interface BlinkIdWasmModule extends BlinkIdBindings, EmscriptenModule {}
  *
  * @ignore
  */
-export interface BlinkIdBindings {
-  createScanningSession: (
-    sessionSettings: BlinkIdSessionSettings,
-    userId: string,
-  ) => BlinkIdScanningSession;
-  initializeWithLicenseKey: (
-    licenceKey: string,
-    userId: string,
-    allowHelloMessage: boolean,
-  ) => LicenseUnlockResult;
-  submitServerPermission: (
-    serverPermission: string,
-  ) => ServerPermissionSubmitError | undefined;
-  getActiveLicenseTokenInfo: () => LicenseUnlockResult;
-  setPingProxyUrl: (url: string) => void;
-  initializeSdk: (userId: string) => void;
-  terminateSdk: () => void;
-  sendPinglets: () => void;
-  arePingRequestsInProgress: () => boolean;
-  queuePinglet: (
-    data: string,
-    schemaName: string,
-    schemaVersion: string,
-    sessionNumber: number,
-  ) => void;
-  isPingEnabled: () => boolean;
-}
+export interface BlinkIdBindings extends WasmBindings<
+  BlinkIdSessionSettings,
+  BlinkIdScanningSession
+> {}