+++ packages/worker-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @microblink/worker-common
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/wasm-common@1.0.1
+
 ## 1.0.0
 
 ### Major Changes